Added fixed threshold
Added the ability to use a fixed threshold method by specifying the threshold value. This is useful for scripting and if you want to apply a consistent threshold value on images with different colour deconvolution vectors.

For this, the old `thresholdMethod` variable under parameters is renamed to `threshold`. If a number is set, it will use that number as the fixed threshold value. If a string is set, it will check if the string is a valid auto threshold method, and if so will calculate a threshold value using the auto threshold.